refactor(stackbuild): split into per-stack files
Split the 1500+ line stackbuild.go into separate files for better
maintainability:

- stackbuild.go: shared types, interfaces, MetaStack, highlevelBuilder
- ruby.go: RubyStack implementation
- python.go: PythonStack implementation
- node.go: NodeStack implementation
- bun.go: BunStack implementation
- golang.go: GoStack implementation
- rust.go: RustStack implementation

No functional changes - pure code reorganization.
